scripts/requirements.txt
# Google API
google-api-python-client==2.84.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0

# Supabase
supabase

# .env file support
python-dotenv

# Progress bar
tqdm